==> notebooks/3_Manual_additions.ipynb <==
# Manually adding cities

### This notebook will be used to add the capital cities that are missing data and cannot be added using our data extraction functions to our city_dataset.

1. Check which cities cannot be obtained with our Numbeo function (for socio-economic data).
2. If there is missing data, we unfortunately cannot move forward with the missing.
3. Check which of the new cities is present in climatedata.eu.
4. For the cities that have missing data, we will be able to manually fill these in as there are many data sources for weather online.
5. Lastly we need to add the latitude and longitude for each city.

import pandas as pd
import numpy as np

city_data = pd.read_json('../data/Combined_data.json')
city_data.head()

capitals = ['Amsterdam', 'Andorra la Vella', 'Athens', 'Belgrade', 'Berlin', 'Bern', 
            'Bratislava', 'Brussels', 'Bucharest', 'Budapest', 'Chisinau', 'Copenhagen', 
            'Dublin', 'Helsinki', 'Kyiv', 'Lisbon', 'Ljubljana', 'London', 'Luxembourg', 
            'Madrid', 'Minsk', 'Monaco', 'Moscow', 'Nicosia', 'Nuuk', 'Oslo', 'Paris', 
            'Podgorica', 'Prague', 'Pristina', 'Reykjavik', 'Riga', 'Rome', 'San Marino', 
            'Sarajevo', 'Skopje', 'Sofia', 'Stockholm', 'Tallinn', 'Tirana', 'Vaduz', 
            'Valletta', 'Vatican City','Vienna', 'Vilnius', 'Warsaw', 'Zagreb']

Checking which cities are still missing from the city_dataset

city_data_list = list(city_data.city)
missing_capitals = []

for capital in capitals:
    if capital not in city_data_list:
        missing_capitals.append(capital)


missing_capitals

### 1.1 Getting the socio-economic data

def get_city_info(city_list) -> list:
    
    # create empty database
    df = pd.DataFrame()
    errors = []
    
    # iterate over list of cities to obtain table data from Numbeo
    for city in city_list:
        try:
            table = pd.read_html(f'https://www.numbeo.com/quality-of-life/in/{city}')[3]
            table = (table.assign(city=city)
                      .drop(columns=[2], axis=1)
                      .rename(columns={0:'category', 1:'numeral'})
                      .drop([table.index[8]]))
            df = df.append(table)
        
        except:
            errors.append(city)
    
    df.reset_index(inplace=True, drop=True)

    return df, errors

capitals_sc = get_city_info(missing_capitals)[0]
errors = get_city_info(missing_capitals)[1]

capitals_sc

### 1.2 The cities in the error list will have to be checked manually.

errors

We'll need to manually add the data for these cities, but if they are not present in Numbeo, they will have to be dropped.

- Andorra la Vella (deleted, missing climate index, missing quality of life index, was an error because Numbeo has the word hyphenated)
- Bern (added, html table at a the wrong index)
- Kyiv (added, spelled 'Kiev' in Numbeo)
- San Marino (deleted, no data in Numbeo)
- Vaduz (deleted, no data in Numbeo)
- Vatican City (deleted, no data in Numbeo)

Adding 'Bern' by updating the function to take the table at the new index.

def manual_city_info(city_list) -> list:
    
    # create empty database
    df = pd.DataFrame()
    errors = []
    
    # iterate over list of cities to obtain table data from Numbeo
    for city in city_list:
        try:
            table = pd.read_html(f'https://www.numbeo.com/quality-of-life/in/{city}')[4]
            table = (table.assign(city=city)
                      .drop(columns=[2], axis=1)
                      .rename(columns={0:'category', 1:'numeral'})
                      .drop([table.index[8]]))
            df = df.append(table)
        
        except:
            errors.append(city)
    
    df.reset_index(inplace=True, drop=True)

    return df, errors

Bern_sc =  manual_city_info(errors)[0]

Bern_sc
