==> notebooks/4_Scraping_climatedata.ipynb <==
# Seasonal weather data for major EU cities

- First we will scrape https://climatedata.eu/ for the cities in our defined list.
- Because climatedata uses special codes in their urls, we will define a dictionary for our cities
- After scraping the data we can subset for seasonal avg temperatures, precipitation and sunshine

import pandas as pd
import lxml

## 1. Getting the data
### 1.1 First, we define a dictionary to hold the url codes for each city

city_codes = {'Amsterdam':'nlxx0002', 'Athens':'grxx0004', 'Belgrade':'srxx0005', 'Berlin':'gmxx0007',
              'Bern':'szxx0006', 'Bratislava':'loxx0001', 'Brussels':'bexx0005', 'Bucharest':'roxx0003',
              'Budapest':'huxx0002', 'Copenhagen':'daxx0009', 'Dublin':'eixx0014', 'Helsinki':'fixx0002', 
              'Lisbon':'poxx0039', 'Ljubljana':'sixx0002', 'London':'ukxx0085', 'Luxembourg':'bexx0027', 
              'Madrid':'spxx0050', 'Monaco':'mnxx0001', 'Nicosia':'cyxx0005', 'Oslo':'noxx0029', 
              'Paris':'frxx0076', 'Prague':'ezxx0012', 'Reykjavik':'icxx0002', 'Riga':'lgxx0004', 
              'Rome':'itxx0067', 'Sarajevo':'bkxx0004', 'Skopje':'mkxx0001', 'Sofia':'buxx0005', 
              'Stockholm':'swxx0031', 'Tallinn':'enxx0004', 'Tirana':'alxx0002', 'Valletta':'mtxx0001', 
              'Vienna':'auxx0025', 'Vilnius':'lhxx0005', 'Warsaw':'plxx0028', 'Zagreb':'hrxx0005'}

**Try to scrape some cities first before creating a function to scrape the whole list.**

amsterdam1 = pd.read_html('https://www.climatedata.eu/climate.php?loc=nlxx0002&lang=en')[2]
amsterdam1

amsterdam2 = pd.read_html('https://www.climatedata.eu/climate.php?loc=nlxx0002&lang=en')[3]
amsterdam2

athens1 = pd.read_html('https://www.climatedata.eu/climate.php?loc=grxx0004&lang=en')[2]
athens1

athens2 = pd.read_html('https://www.climatedata.eu/climate.php?loc=grxx0004&lang=en')[3]
athens2

### 1.2 Now we'll write a function to gather all the cities's data in one go

test_dict = {'Amsterdam':'nlxx0002', 'Athens':'grxx0004', 'Vienna':'auxx0025'}

def get_data(city_dict) -> dict:
    df = pd.DataFrame()
    errors = []

    for city, code in city_dict.items():
        try:
            # get first table (from jan to june)
            table1 = pd.read_html(f'https://www.climatedata.eu/climate.php?loc={code}&lang=en')[2]
        
            # get second table (from july to dec)
            table2 = pd.read_html(f'https://www.climatedata.eu/climate.php?loc={code}&lang=en')[3]
            
            # merge table1 and table2
            table_combo = table1.merge(table2)
            table_combo = table_combo.assign(city=city)
            
            # assign tables to df
            df = df.append(table_combo)
        
        except:
            errors.append(city)

    return df
    

test_weather = get_data(test_dict)

city_weather = get_data(city_codes)

city_weather.head(20)

## 2. Now we have our dataframe, it's time to inspect and subset
### 2.1 checking dtypes and updating colnames

city_weather.dtypes

city_weather = city_weather.rename(columns={'Unnamed: 0':'weather'})

city_weather.head()

### 2.2 Grouping months into seasons

#### Keep in mind that the data is not standardised (e.g we have temperatures in Celius, number of mm, days and hours)

test_weather

round((test_weather["Dec"] + test_weather["Jan"] + test_weather["Feb"]) / 3)

test_weather.head()

def get_season_avg(df):
    season_dict = {"winter": ["Dec", "Jan", "Feb"],
                   "spring": ["Mar", "Apr", "May"],
                   "summer": ["June", "July", "Aug"],
                   "autumn": ["Sep", "Oct", "Nov"],
                   }
    
    for season, month in season_dict.items():
        # Get winter averages
        df[season] = round((df[month[0]] + df[month[1]] + df[month[2]]) / 3)
        df = df.drop(df[month], axis=1)

    return df 

get_season_avg(test_weather)

city_seasonal_avg = get_season_avg(city_weather)

city_seasonal_avg

### 2.3 Now to pivot the table so each city only has one row

test_df = city_seasonal_avg.loc[city_seasonal_avg['city'] == 'Amsterdam']

test_df

colnames = ['autumn_prec_mm', 'autumn_high', 'autumn_low', 'autumn_prec_days', 'autumn_sun_hrs',
            'spring_prec_mm', 'spring_high', 'spring_low', 'spring_prec_days', 'spring_sun_hrs',
            'summer_prec_mm', 'summer_high', 'summer_low', 'summer_prec_days', 'summer_sun_hrs',
            'winter_prec_mm', 'winter_high', 'winter_low', 'winter_prec_days', 'winter_sun_hrs']

test_pivot = test_df.pivot_table(index='city', columns=['weather'])
test_pivot

test_pivot.columns = [' '.join(col).strip() for col in test_pivot.columns.values]

test_pivot.columns = [colnames]

test_pivot.reset_index(inplace=True)
test_pivot

### 2.4 Putting it all into a function 

def transform_df(df):
    df = df.rename(columns={'Unnamed: 0':'weather'})
    colnames = ['autumn_prec_mm', 'autumn_high', 'autumn_low', 'autumn_prec_days', 'autumn_sun_hrs',
                'spring_prec_mm', 'spring_high', 'spring_low', 'spring_prec_days', 'spring_sun_hrs',
                'summer_prec_mm', 'summer_high', 'summer_low', 'summer_prec_days', 'summer_sun_hrs',
                'winter_prec_mm', 'winter_high', 'winter_low', 'winter_prec_days', 'winter_sun_hrs',
               ]
    
    df = df.pivot_table(index='city', columns=['weather'])
    df.columns = df.columns.get_level_values(1)
    df.columns = [colnames]
    df.reset_index(inplace=True)
    
    return df

transform_df(city_seasonal_avg).head()

def to_seasonal(city_dict) -> dict:
    df = get_data(city_dict)
    df = get_season_avg(df)
    df = transform_df(df)
    
    return df
    

weather_final = to_seasonal(city_codes)

weather_final

weather_final.columns = weather_final.columns.get_level_values(0)
weather_final.columns

## 3. Exporting to json

weather_final.to_json('../data/Seasonal_weather_data.json')
