==> notebooks/5_Scraping_numbeo.ipynb <==
# Obtaining socio-economic data
#### Using pandas read_html to collect table data from https://www.numbeo.com/quality-of-life/in/Berlin

import pandas as pd
import lxml

## 1. Getting city data
### 1.1 Define list of the European capital cities

city_list = ['Amsterdam', 'Andorra-La-Vella', 'Athens', 'Belgrade', 'Berlin', 'Bern', 
             'Bratislava', 'Brussels', 'Bucharest', 'Budapest', 'Chisinau', 'Copenhagen', 
             'Dublin', 'Helsinki', 'Kiev', 'Lisbon', 'Ljubljana', 'London', 'Luxembourg', 
             'Madrid', 'Minsk', 'Monaco', 'Moscow', 'Nicosia', 'Nuuk', 'Oslo', 'Paris', 
             'Podgorica', 'Prague', 'Reykjavik', 'Riga', 'Rome', 'San-Marino', 'Sarajevo', 
             'Skopje', 'Sofia', 'Stockholm', 'Tallinn', 'Tirana', 'Vaduz', 'Valletta', 'Vatican-City', 
             'Vienna', 'Vilnius', 'Warsaw', 'Zagreb']

### 1.2 Creating a function to obtain data and put it into a dataframe for processing and modelling

test_list = ['Vienna', 'Berlin']

def get_city_info(city_list) -> list:
    
    # create empty database
    df = pd.DataFrame()
    errors = []
    
    # iterate over list of cities to obtain table data from Numbeo
    for city in city_list:
        try:
            table = pd.read_html(f'https://www.numbeo.com/quality-of-life/in/{city}')[3]
            table = (table.assign(city=city)
                      .rename(columns={0:'category', 1:'numeral', 2:'level'})
                      .drop([table.index[8]]))
            df = df.append(table)
        
        except:
            errors.append(city)
    
    df.reset_index(inplace=True, drop=True)
    
    # df = df.pivot(index='city', columns='category', values=['numeral', 'level'])
    
    # colnames = ['climate', 'cost_of_living', 'health_care', 'pollution', 'property_income_ratio', 'purchasing_power',
                #'safety', 'traffic_time', 'quality_of_life', 'climate_level', 'cost_of_living_level',
                #'health_care_level', 'pollution_level', 'property_income_ratio_level', 'purchasing_power_level',
                #'safety_level', 'traffic_time_level', 'quality_of_life_level']
                
    #df.columns = colnames
    #df = df.reset_index()
        
    return df, errors
 
    

get_city_info(test_list)

## 2. Final data-set
### 2.1 Checking for cities not covered by Numbeo

#get_city_info(city_list)

**After running the function on the whole list San Marino, Vaduz and Vatican City appeared in the error list**

drop_cities = ['San-Marino', 'Vaduz', 'Vatican-City']

update_cities = [city for city in city_list if city not in drop_cities]

print(update_cities)

### 2.2 Now rerunning function with updated list

# get_city_info(update_cities)

### 2.3 There seem to be no errors coming up, now to rewrite the function, so it returns a data-set again

def get_city(city_list) -> list:
    
    # create empty database
    df = pd.DataFrame()
    
    # iterate over list of cities to obtain table data from Numbeo    
    # removing index position 8, because it refers to a border
    for city in city_list:
        table = pd.read_html(f'https://www.numbeo.com/quality-of-life/in/{city}')[3]
        table = (table.assign(city=city)
                      .rename(columns={0:'category', 1:'numeral', 2:'level'}))
        df = df.append(table)
    
    df.reset_index(inplace=True, drop=True)
  
    return df

#all_cities = 
get_city(test_list)

# all_cities

### 2.4 Checking values of Index and Unnamed: 1 columns
There seems to be a city whose data was pushed back by one.

# all_cities['Index'].value_counts()

# all_cities['Unnamed: 1'].value_counts()

# all_cities.loc[all_cities['Unnamed: 1'] == 209.07]

The data for the city of Bern seems to have been scraped incorrectly. Let's check what data Numbeo actually has available. 

pd.read_html(f'https://www.numbeo.com/quality-of-life/in/Bern')[3]

Perhaps we are reading the wrong table from Numbeo, let's try the others.

pd.read_html(f'https://www.numbeo.com/quality-of-life/in/Bern')[4]

It seems as though there is not enough user information for Bern and there is a notice table in the html
pushing it to index position 4. We will delete Bern from the list and add it later manually to the list.

update_cities.remove('Bern')

print(update_cities)

#final_cities = get_city(update_cities)

def get_city(city_list) -> list:
    
    # create empty database
    df = pd.DataFrame()
    
    # iterate over list of cities to obtain table data from Numbeo    
    # removing index position 8, because it refers to a border
    for city in city_list:
        table = pd.read_html(f'https://www.numbeo.com/quality-of-life/in/{city}')[3]
        table = (table.assign(city=city)
                      .rename(columns={0:'category', 1:'numeral', 2:'level'})
                      .drop([table.index[8]]))
        df = df.append(table)
    
    df.reset_index(inplace=True, drop=True)
  
    return df

final_cities = get_city(update_cities)

final_cities.head(20)

**Add Bern manually**

bern = pd.read_html(f'https://www.numbeo.com/quality-of-life/in/Bern')[4]
bern = (bern.assign(city='Bern')
        .rename(columns={0:'category', 1:'numeral', 2:'level'})
        .drop([bern.index[8]]))

bern

final_cities = final_cities.append(bern)
final_cities.reset_index(inplace=True, drop=True)

final_cities[final_cities['city'] == 'Bern']

### 2.5 Now to pivot the table so each city only has one row

def pivot_cities(df):

    df = df.pivot(index='city', columns='category', values=['numeral', 'level'])
    
    colnames = ['climate', 'cost_of_living', 'health_care', 'pollution', 'property_income_ratio', 'purchasing_power',
                'safety', 'traffic_time', 'quality_of_life', 'climate_level', 'cost_of_living_level',
                'health_care_level', 'pollution_level', 'property_income_ratio_level', 'purchasing_power_level',
                'safety_level', 'traffic_time_level', 'quality_of_life_level']
    
    df.columns = colnames
    df = df.reset_index()
    
    return df

final_cities = pivot_cities(final_cities)

final_cities.head(20)

## 3. Data cleaning on final table

### 3.1 replacing the one '?' with NaN

import numpy as np

final_cities = final_cities.replace('?', np.nan)
final_cities

### 3.2 Fixing the data types

final_cities.dtypes

cols_to_float = ['climate', 'cost_of_living', 'health_care', 'pollution',
                 'property_income_ratio', 'purchasing_power', 'safety',
                 'traffic_time']

final_cities[cols_to_float] = final_cities[cols_to_float].apply(pd.to_numeric)

final_cities.dtypes

## 4. Exporting to json format

final_cities.to_json('../data/Socio_economic_data.json')
