==> notebooks/6_Merging_all_data.ipynb <==
# Combinging the data for EDA

This notebook is no longer used to export to the city_dataset, it is solely for reference as to how we originally merged the data. The new method for data extraction and cleaning is in the notebook 'Data_sammler' (folder 'new_features') and is still in-progress.

## 1. Importing and merging

### 1.1 Getting the json files and converting them to dataframes for a quick overview

import pandas as pd

Due to limited functionality with the Foursquare API, we will no longer be using venue data, until we find a better source.

# Importing Foursquare venues data
venues_data = pd.read_json('../data/Foursquare_venues_data.json')
venues_data.head()

# Importing seasonal weather data
weather_data = pd.read_json('../data/Seasonal_weather_data.json', orient='columns')
weather_data.head()

# Importing socio-economic data
soc_econ_data = pd.read_json('../data/Socio_economic_data.json')
soc_econ_data.head()

### 1.2 Checking dtypes to make sure all columns are correct

venues_data.dtypes

weather_data.dtypes

soc_econ_data.dtypes

## 2. Putting it all together
### 2.1 Using a merge on "city" with inner join, because right now we do not mind losing a few cities for our testing data.

all_data = venues_data.merge(weather_data).merge(soc_econ_data)

all_data.head()

print(all_data.columns)

### 2.2 Changing Venues data back to 0's instead of NaN

# We have 268 columns that need zeroes instead of NaNs
venues_data.shape

venue_cols = venues_data.columns

all_data[venue_cols] = all_data[venue_cols].fillna(0)

all_data.head()

### 2.3 Exporting as json

all_data.shape

all_data.dtypes

# all_data.to_json('../data/Combined_data.json')
